Parte de devoluciones de prestamos completada* Requiere testing

+++ src/gestorAplicacion/productos/Producto.java
@@ -155,7 +155,7 @@ public abstract class Producto implements Serializable, Cloneable,Comparable<Pro
 	public String toStringPrestable() {
 		return "COD: " + codigo + " | " +
 				"NOMBRE: " + nombre + " | " +
-				"VALOR PRESTAMO: $ " + (valor * 0.10) + " | " +
+				"VALOR PRESTAMO: $ " + (int) (valor * 0.10) + " | " +
 				"CANT PRESTABLE: " + cantidad;
 	}
